initial commit
This is a simple multi-mod waveform generator. Right now I have
demonstrated that I can actually use the GR sinks inside of a view
created with qt4-designer and pyuic4. Nothing too novel, but the
intention is that this turns in to an end-user application that can take
any signal source, modulate it with any GR modulation, then send it to
any signal sink as well as showing a snapshot in time and showing
freq/waterfalls. This is both a simple useful app and a proof of concept